thien push bill manager staff

+++ Rentify/app/src/main/java/com/rentify/user/app/model/Model/User.kt
@@ -81,7 +81,7 @@ data class LandlordResponse(
 
 data class BankAccount(
     val bank_name: String,
-    val bank_number: Int,
+    val bank_number: Long,
     val qr_bank: List<String>,
     val username: String
 )
